Note timing in double

+++ src/main/java/com/DAWIntegration/Satbify/module/Note.java
@@ -1,4 +1,4 @@
 package com.DAWIntegration.Satbify.module;
 
-public record Note(int reaperTrack, int pitch, int start, int end) {}
+public record Note(int reaperTrack, int pitch, double start, double end) {}
